@@ -13,7 +13,7 @@ public class ResetDependenciesCommandSettings : ConfigureDependenciesCommandSett
 {
     [Description( "The list of dependencies (given by name by or position in the dependency list) to configure" )]
     [CommandArgument( 1, "[dependencies]" )]
-    public string[] Dependencies { get; protected set; } = Array.Empty<string>();
+    public string[] Dependencies { get; protected set; } = [];
 
     [Description( "Specifies that all dependencies must be configured" )]
     [CommandOption( "--all" )]
