@@ -45,13 +45,13 @@ public static partial class MetalamaDependencies
 
         public static ProductFamily Family { get; } = new( _projectName, "2024.2", DevelopmentDependencies.Family, PostSharpDependencies.V2024_1.Family )
         {
-            UpstreamProductFamily = V2024_1.Family
+            DockerBaseImage = "mcr.microsoft.com/dotnet/sdk:8.0.203-jammy", UpstreamProductFamily = V2024_1.Family
 
             // DownstreamProductFamily = V2024_3.Family
         };
 
         public static DependencyDefinition MetalamaBackstage { get; } = new MetalamaDependencyDefinition( "Metalama.Backstage", VcsProvider.GitHub );
-        
+
         public static DependencyDefinition Consolidated { get; } = new MetalamaDependencyDefinition(
             "Consolidated",
             VcsProvider.AzureDevOps,
@@ -66,7 +66,8 @@ public static partial class MetalamaDependencies
             EngineeringDirectory = "eng-Metalama", PrivateArtifactsDirectory = Path.Combine( "artifacts", "packages", "$(MSSBuildConfiguration)", "Shipping" )
         };
 
-        public static DependencyDefinition MetalamaFrameworkRunTime { get; } = new MetalamaDependencyDefinition( "Metalama.Framework.RunTime", VcsProvider.GitHub );
+        public static DependencyDefinition MetalamaFrameworkRunTime { get; } =
+            new MetalamaDependencyDefinition( "Metalama.Framework.RunTime", VcsProvider.GitHub );
 
         public static DependencyDefinition MetalamaFrameworkPrivate { get; } = new MetalamaDependencyDefinition(
             "Metalama.Framework.Private",
@@ -85,7 +86,7 @@ public static partial class MetalamaDependencies
 
         public static DependencyDefinition MetalamaSamples { get; } =
             new MetalamaDependencyDefinition( "Metalama.Samples", VcsProvider.GitHub ) { CodeStyle = "Metalama.Samples" };
-        
+
         public static DependencyDefinition TimelessDotNetEngineer { get; } =
             new MetalamaDependencyDefinition( "TimelessDotNetEngineer", VcsProvider.GitHub ) { CodeStyle = "Metalama.Samples" };
 
