@@ -76,14 +76,28 @@ namespace PostSharp.Engineering.BuildTools.Utilities
                 workingDirectory,
                 ConsoleHelper.CancellationToken,
                 out exitCode,
-                s =>
+                HandleErrorData,
+                HandleOutputData,
+                options );
+
+            void HandleErrorData( string s )
+            {
+                if ( options.FilterOutput )
                 {
                     if ( !string.IsNullOrWhiteSpace( s ) )
                     {
                         console.WriteMessage( s );
                     }
-                },
-                s =>
+                }
+                else
+                {
+                    Console.Error.WriteLine( s );
+                }
+            }
+
+            void HandleOutputData( string s )
+            {
+                if ( options.FilterOutput )
                 {
                     if ( !string.IsNullOrWhiteSpace( s ) )
                     {
@@ -122,8 +136,12 @@ namespace PostSharp.Engineering.BuildTools.Utilities
                             console.WriteMessage( s );
                         }
                     }
-                },
-                options );
+                }
+                else
+                {
+                    Console.Out.WriteLine( s );
+                }
+            }
         }
 
         // #16205 We don't allow cancellation here because there's no other working way to wait for a process exit
@@ -228,21 +246,20 @@ namespace PostSharp.Engineering.BuildTools.Utilities
                     startInfo.Environment[blockedEnvironmentVariable] = null;
                 }
 
-                // Filters process output where matching RegEx value indicates process failure.
-                void FilterProcessOutput( string output )
-                {
-                    if ( options.Retry is { Regex: { } } && options.Retry.Regex.IsMatch( output ) )
-                    {
-                        processShouldRetry = true;
-                    }
-                }
-
-                Path.GetFileName( fileName );
                 Process process = new() { StartInfo = startInfo };
 
                 using ( ManualResetEvent stdErrorClosed = new( false ) )
                 using ( ManualResetEvent stdOutClosed = new( false ) )
                 {
+                    // Filters process output where matching RegEx value indicates process failure.
+                    void FilterProcessOutput( string output )
+                    {
+                        if ( options.Retry is { Regex: { } } && options.Retry.Regex.IsMatch( output ) )
+                        {
+                            processShouldRetry = true;
+                        }
+                    }
+
                     process.ErrorDataReceived += ( _, args ) =>
                     {
                         try
@@ -255,7 +272,7 @@ namespace PostSharp.Engineering.BuildTools.Utilities
                             {
                                 FilterProcessOutput( args.Data );
 
-                                handleErrorData( args.Data );
+                                handleErrorData!( args.Data );
                             }
                         }
                         catch ( Exception e )
@@ -276,7 +293,7 @@ namespace PostSharp.Engineering.BuildTools.Utilities
                             {
                                 FilterProcessOutput( args.Data );
 
-                                handleOutputData( args.Data );
+                                handleOutputData!( args.Data );
                             }
                         }
                         catch ( Exception e )
@@ -317,6 +334,7 @@ namespace PostSharp.Engineering.BuildTools.Utilities
                         {
                             process.BeginErrorReadLine();
                             process.BeginOutputReadLine();
+
                             process.WaitForExit();
                         }
                         else
